# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * J&K Premium Casino API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  username: string;
  balance: number;
  createdAt: string;
}

export interface UserRegisterInput {
  /** @minLength 3 */
  username: string;
}

export interface BalanceUpdate {
  /** Amount to add (positive) or subtract (negative) */
  amount: number;
}

export interface Transfer {
  id: string;
  fromUserId: string;
  toUserId: string;
  fromUsername?: string;
  toUsername?: string;
  amount: number;
  createdAt: string;
}

export interface TransferInput {
  fromUserId: string;
  toUserId: string;
  /** @minimum 1 */
  amount: number;
}

export interface GameRound {
  id: string;
  userId: string;
  betAmount: number;
  multiplier: number;
  /** @nullable */
  cashoutMultiplier: number | null;
  won: boolean;
  /** @nullable */
  profit?: number | null;
  createdAt: string;
}

export interface GameRoundInput {
  userId: string;
  betAmount: number;
  multiplier: number;
  /** @nullable */
  cashoutMultiplier?: number | null;
  won: boolean;
  /** @nullable */
  profit?: number | null;
}

export interface GameStats {
  totalRounds: number;
  totalWon: number;
  totalLost: number;
  biggestWin: number;
  totalProfit: number;
}