# lib/api-zod/src/generated/types/balanceUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * J&K Premium Casino API
 * OpenAPI spec version: 0.1.0
 */

export interface BalanceUpdate {
  /** Amount to add (positive) or subtract (negative) */
  amount: number;
}