# lib/api-zod/src/generated/types/gameRoundInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * J&K Premium Casino API
 * OpenAPI spec version: 0.1.0
 */

export interface GameRoundInput {
  userId: string;
  betAmount: number;
  multiplier: number;
  /** @nullable */
  cashoutMultiplier?: number | null;
  won: boolean;
  /** @nullable */
  profit?: number | null;
}