# lib/api-zod/src/generated/types/gameStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * J&K Premium Casino API
 * OpenAPI spec version: 0.1.0
 */

export interface GameStats {
  totalRounds: number;
  totalWon: number;
  totalLost: number;
  biggestWin: number;
  totalProfit: number;
}