# lib/api-zod/src/generated/types/transfer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * J&K Premium Casino API
 * OpenAPI spec version: 0.1.0
 */

export interface Transfer {
  id: string;
  fromUserId: string;
  toUserId: string;
  fromUsername?: string;
  toUsername?: string;
  amount: number;
  createdAt: string;
}