# lib/api-zod/src/generated/types/transferInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * J&K Premium Casino API
 * OpenAPI spec version: 0.1.0
 */

export interface TransferInput {
  fromUserId: string;
  toUserId: string;
  /** @minimum 1 */
  amount: number;
}