# lib/api-zod/src/generated/types/userRegisterInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * J&K Premium Casino API
 * OpenAPI spec version: 0.1.0
 */

export interface UserRegisterInput {
  /** @minLength 3 */
  username: string;
}